requirements.txt
mysql-connector-python
fastapi
uvicorn
starlette
python-dotenv
pydantic